Тестирование/django-locallibrary-tutorial-master/requirements.txt
dj-database-url==0.5.0
Django==3.1.2
gunicorn==20.0.4
psycopg2-binary==2.8.6
wheel==0.35.1
whitenoise==5.2.0
djangorestframework
parameterized
